+++ src/components/Popover/Popover.vue
@@ -5,7 +5,7 @@
     type="black"
     position="top-left"
     :data-testid="props.dataTestid"
-    :is-can-show="!!props.tooltip && !state.isShow"
+    :is-can-show="!!props.tooltip && !state.isShow && props.options.length > 0"
   >
     <div
       class="popover-yui-kit"
@@ -53,7 +53,6 @@ import { IPopoverOption, IPopoverProps } from './interface/interface';
 import Icon from '@/components/Icon/Icon.vue';
 import { IconNameEnum } from '../Icon/enum/enum';
 import { vOnClickOutside } from '@vueuse/components';
-import type { OnClickOutsideHandler } from '@vueuse/core';
 import Tooltip from '@/components/Tooltip/Tooltip.vue';
 
 interface IPopoverState {
@@ -67,6 +66,10 @@ const props = withDefaults(defineProps<IPopoverProps>(), {
   dataTestid: 'Popover'
 });
 
+const emits = defineEmits<{
+  (e: 'close'): void;
+}>();
+
 const state = reactive<IPopoverState>({
   isShow: false
 });
@@ -79,7 +82,8 @@ const currentRef = ref<HTMLElement | null>(null);
  */
 const classesFilter = computed(() => ({
   'popover-yui-kit__wrapper': true,
-  'active-yui-kit': state.isShow
+  'active-yui-kit': state.isShow,
+  'empty-yui-kit': props.options.length === 0
 }));
 
 /**
@@ -87,8 +91,9 @@ const classesFilter = computed(() => ({
  */
 const toggleShow = () => (state.isShow = !state.isShow);
 
-const closeShow: OnClickOutsideHandler = (): void => {
+const closeShow = (): void => {
   state.isShow = false;
+  emits('close');
 };
 
 const handleClick = (item: IPopoverOption): void => {
@@ -104,16 +109,20 @@ const updateDropdownPosition = () => {
   }
 };
 
-onMounted(updateDropdownPosition);
-
 watch(() => state.isShow, updateDropdownPosition);
 
+watch(
+  () => props.isShow,
+  () => (state.isShow = props.isShow)
+);
+
 onMounted(() => {
-  window.addEventListener('scroll', () => (state.isShow = false), true);
+  updateDropdownPosition();
+  window.addEventListener('scroll', closeShow, true);
 });
 
 onUnmounted(() => {
-  window.removeEventListener('scroll', () => (state.isShow = false), true);
+  window.removeEventListener('scroll', closeShow, true);
 });
 </script>
 
@@ -134,6 +143,13 @@ onUnmounted(() => {
     &:hover,
     &.active-yui-kit {
       color: var(--border-blue);
+      &.empty-yui-kit {
+        color: var(--popover-icon-color, var(--grey4));
+      }
+    }
+
+    &.empty-yui-kit {
+      cursor: default;
     }
   }